backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
jinja2
python-multipart
python-dotenv
openai
weasyprint
mysql-connector-python
psycopg2-binary
